clients/pledge: Move components into one module
